Switched Calendar2_Models_Calendar2->fetchAllForPeriod to accept multiple users.
The $users parameter (formerly $users) can now also be an array, meaning that
all events that any of the given users participates in should be returned. The
current behaviour when passing null or a single id is not changed.

We don't want the server to have separate 'dayViewSelect', 'weekViewSelect',
etc. methods like the client currently has.