Use Zend_Application
This commit introduces a Bootstrap.php, which handles most of the initialization
tasks formerly done in Phprojekt->_initialize(). The Tasks have been divided
into different resources, which depend on each other throught the bootstrap()
function provided by the Zend Bootstrapping mechanism.

It also adds an application.ini, which covers the static initialization parts
like default timezone, all static paths (the per-module directories are dynamic
and have to be set in code) and the routes.

The database is not created by Phprojekt anymore, as we are now using the Zend
Resource. This forces us to change some configuration keys, which will be fixed
in the next commit.

Phprojekt->_get{Config,Db,Log,Cache} just reach into the resource registry
provided by bootstrap.